+++ flake.nix
@@ -6,7 +6,8 @@
   inputs = {
   	nixpkgs.url = "nixpkgs/nixos-unstable";
 	#hyprland.url = "github:hyprwm/Hyprland"; # hyprland development
-	distro-grub-themes.url = "github:AdisonCavani/distro-grub-themes"; 
+	distro-grub-themes.url = "github:AdisonCavani/distro-grub-themes";
+	ags.url = "github:aylur/ags/v1"; # aylurs-gtk-shell-v1
   	};
 
   outputs =